Output callback chunk entries based on #tag{} records
Merge back into the OTP tree caused the eep48_SUITE:cb_anno test to fail.
The reason was that callbacks were not present in the XML EDoc format,
but they were present in #entry{} and #tag{} representation.
Since we use this representation in the chunk layout already,
using it is the easiest way to get chunk output back.