Adjust default/private/hidden doc visibility in chunks
Prior to these adjustments when using ExDoc to generate docs for
Erlang projects modules were required to have some top-level docs to be
listed in the generated HTML. They could still have function-level docs,
but the module wouldn't be listed anyway, so these docs wouldn't be accessible.

With these changes a module will be listed even if it does not have
top-level docs, so that its function docs are available in the generated HTML.
This means a module has to be explicitly marked with EDoc @private or
@hidden not to be listed.

See also https://github.com/elixir-lang/ex_doc/issues/1393.