docs: Add tools to generate markdown from xml
The tools take application/html+erlang EEP-48 style documentation
and convert it to EEP-59 style markdown. All erl_docgen types of
documentation files are supported.

This commit also adds some rudimentory support for generating HTML
docs using ex_doc. The support is basically there to test that
the XML to Markdown convertion works as intended and is not meant
for production use.