Replace "on the form" with "of the form"
"On the form" is used when something is written on a paper or
electronic form, for example on a tax form.

To express that something has the shape or structure of something
else, the correct expression is "of the form" or "in the form".

Thanks to Richard O'Keefe for pointing this out. See
https://erlangforums.com/t/otp-28-0-rc1-released/4482/3.